Include `=` when matching aliases' definitions
This fixes the scenario where an alias exists who's name starts with
another alias' name:

alias foo_bar=1
alias foo=2

When editing the alias 'foo', the curser will be placed on the first
alias instead of the second.
This commit includes the '=' in the match, so 'we foo' will search for
'foo=' and will thus not match the 'foo_bar' alias anymore.